chore: use TargetFrameworks for compiling with .NET Standard 2.0 (for Unity)

+++ src/ShaderLab/SharpX.ShaderLab.Primitives/Types/AppDataImg.cs
@@ -16,5 +16,5 @@ public struct AppDataImg
     public Vector4<float> Vertex { get; }
 
     [Name("texcoord")]
-    public Vector2<Half> TexCoord { get; }
+    public Vector2<float> TexCoord { get; }
 }
